+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/api/functions/aggregation/PreAggregateTrigger.java
@@ -24,9 +24,9 @@ public interface PreAggregateTrigger<T> extends Serializable {
 
 	String explain();
 
-	long getMaxCount();
+	int getMaxCount();
 
-	void setMaxCount(long maxCount, int subtaskIndex);
+	void setMaxCount(int maxCount, int subtaskIndex);
 
 	PreAggregateStrategy getPreAggregateStrategy();
 }
